requirements.txt
python-dotenv
schedule
requests
pyyaml